Further Task Editing
bugfix: Fixed bugs relating to task editing,
Tasks now can be properly saved and edited.

feat: Task Info Popup Reusability
The TodoPopup is reuable and only requires two paramaters to be specified when implementing the xaml. This is the "RunWhenSave" and "Todo" parameter. RunWhenSave is a bindable ICommand which will run when the save button is pressed. Intended to be used for updating the database. The Todo paramater is the Todo object which is to be bound to the popup. Changes made in the popup will be reflected in the bound todo object.

feat: Editing cancel
Cancelling an edit will cause the popup to show the orignal data.

feat: Task popup importance
Tasks now have editable and viewable importance buttons.

feat: Task colour picking
Task popup now has labels for colour. Task's colour can be changed properly.

todo: Task deletion
Allow for tasks to be deleted