Notification Creation + Restoring
feat!: App will run as a service in the background.
The app attaches a broadcast receiver to get the boot intent. On intent recieved the app will run in the background as a service allowing for the resoration of notifications and priority recalculation.

feat: Permission request
The app will request for permissions for sending notifications and starting on boot if not granted.

feat!: Notifications for task creation.
If a task has notification's turned on for them a notification will be added to the db and scheduled at the correct time.

feat: Methods for notification schedule generation
Notifications will automatically be scheduled using the type and notification data.