Automatic Deletion
feat: Automatic Task deletion
Tasks that have been sent to the recycle bin will automatically be deleted after 30 days. This is accomplished using the DeleteModel which stores the time a task was sent to the bin. It contains methods to generate these times and to automatically delete the tasks from the DB after 30 days.